+++ napari/conftest.py
@@ -1,13 +1,9 @@
 import os
-import sys
-import warnings
 from functools import partial
-from typing import List
 
 import numpy as np
 import pytest
 
-from napari import Viewer
 from napari.components import LayerList
 from napari.layers import Image, Labels, Points, Shapes, Vectors
 from napari.plugins._builtins import (
@@ -64,49 +60,6 @@ def pytest_addoption(parser):
         default=False,
         help="run only asynchronous tests",
     )
-
-
-@pytest.fixture
-def qtbot(qtbot):
-    """A modified qtbot fixture that makes sure no widgets have been leaked."""
-    from qtpy.QtWidgets import QApplication
-
-    initial = QApplication.topLevelWidgets()
-    prior_exception = getattr(sys, 'last_value', None)
-
-    yield qtbot
-
-    # if an exception was raised during the test, we should just quit now and
-    # skip looking for leaked widgets.
-    if getattr(sys, 'last_value', None) is not prior_exception:
-        return
-
-    QApplication.processEvents()
-    leaks = set(QApplication.topLevelWidgets()).difference(initial)
-    # still not sure how to clean up some of the remaining vispy
-    # vispy.app.backends._qt.CanvasBackendDesktop widgets...
-    if any([n.__class__.__name__ != 'CanvasBackendDesktop' for n in leaks]):
-        raise AssertionError(f'Widgets leaked!: {leaks}')
-    if leaks:
-        warnings.warn(f'Widgets leaked!: {leaks}')
-
-
-@pytest.fixture(scope="function")
-def make_test_viewer(qtbot, request):
-    viewers: List[Viewer] = []
-
-    def actual_factory(*model_args, viewer_class=Viewer, **model_kwargs):
-        model_kwargs['show'] = model_kwargs.pop(
-            'show', request.config.getoption("--show-viewer")
-        )
-        viewer = viewer_class(*model_args, **model_kwargs)
-        viewers.append(viewer)
-        return viewer
-
-    yield actual_factory
-
-    for viewer in viewers:
-        viewer.close()
 
 
 @pytest.fixture(
