+++ setup.cfg
@@ -111,6 +111,8 @@ dev =
 [options.entry_points]
 console_scripts =
     napari = napari.__main__:main
+pytest11 =
+    napari = napari.utils._testsupport
 
 
 [flake8]
